SearchDetail useBookDetail api 적용
