+++ src/pages/Search/SearchDetail/Header/SearchDetailHeader.module.css
@@ -6,6 +6,7 @@
   display: flex;
   align-items: center;
   width: 21.4375rem;
+  /* height: 3.5rem; */
   justify-content: space-between;
   color: white;
   font-weight: bold;
@@ -17,6 +18,7 @@
 .title {
   flex-grow: 1;
   text-align: center;
+  font-size: 1.35rem;
 }
 
 .icon {
@@ -27,7 +29,9 @@
   justify-content: center;
 }
 
-/* .icon img {
-  max-width: 100%;
-  max-height: 100%;
-} */
+
+
+.icon img {
+  width: 100%;
+  height: 100%;
+}
